Optimize Bruce Lee

+++ catscripts/Entertainment/Actors_and_actresses/Bruce_Lee/BruceLee.sh
@@ -1,16 +1,19 @@
 #!/bin/bash
 
-export KEYWORDS_ACTOR_BRUCELEE="Bruce(| )Lee"
+KEYWORDS_ACTOR_BRUCELEE="Bruce(| )Lee"
 
-egrep -i "$KEYWORDS_ACTOR_BRUCELEE" newpages.txt >> BruceLee.txt
-
-BRUCELEE=`stat --print=%s BruceLee.txt`
-
-if [ $BRUCELEE -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="BruceLee.txt"
-  export CATNAME="Bruce Lee"
-  $CATEGORIZE
-fi
 
-rm BruceLee.txt
+  BRUCELEE=`egrep -i "$KEYWORDS_ACTOR_BRUCELEE" newpages.txt`
+
+  if [ "$BRUCELEE" != "" ];
+  then
+    egrep -i "$KEYWORDS_ACTOR_BRUCELEE" newpages.txt > BruceLee.txt
+    export CATFILE="BruceLee.txt"
+    export CATNAME="Bruce Lee"
+    $CATEGORIZE
+    rm BruceLee.txt
+    unset BRUCELEE
+  fi
+fi
